pyproject.toml
[project]
name = "auto-annotator"
version = "0.1.0"
description = "AI-powered video annotation system for multimodal sports datasets"
readme = "README.md"
authors = [
    { name = "qi7876", email = "97651950+qi7876@users.noreply.github.com" }
]
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "opencv-python>=4.8.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.8.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
